requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
pydantic==2.11.10
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
